+++ pubspec.yaml
@@ -1,18 +1,19 @@
 name: youtube_api
-description: YoutubeAPI is a plugin which directly interects with youtube server. Supports
-               Searching video and playlist.
-version: 0.7.2
+description:
+  YouTubeAPI is a plugin which directly interacts with YouTube server. Supports
+  Searching video and playlist.
+version: 0.8.0
 author: NITISH KUMAR SINGH <nitishk2@gmail.com>
 homepage: https://github.com/nitishk72/Youtube_API
 
 dependencies:
   flutter:
     sdk: flutter
-    
+
   http: ^0.12.0+1
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: ">=2.0.0 <3.0.0"
 
 flutter:
   plugin:
